requirements.txt
certifi
chardet
idna
requests
selenium
urllib3
ipapi
undetected-chromedriver
mysql-connector-python
pytz